Moved config post-processors to their own sub-namespace
