+++ module/Core/src/Config/PostProcessor/BasePathPrefixer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Core\Config;
+namespace Shlinkio\Shlink\Core\Config\PostProcessor;
 
 use function Functional\map;
